requirements.txt
pandas
numpy
matplotlib
requests
python-binance
ta
websocket-client